requirements.txt
flask
flask-sqlalchemy
sqlalchemy
psycopg2-binary
pyyaml
gunicorn
pytest
opencv-python